refactor: organize deployment detail files
