+++ web/features/deployments/detail/components/detail-table-styles.ts
