+++ web/features/deployments/detail/components/detail-table.tsx
