+++ web/features/deployments/detail/settings-tab/access/access-tab.tsx
@@ -1,7 +1,7 @@
 'use client'
 
-import { AccessChannelsSection } from './settings-tab/access/channels-section'
-import { AccessPermissionsSection } from './settings-tab/access/permissions-section'
+import { AccessChannelsSection } from './channels-section'
+import { AccessPermissionsSection } from './permissions-section'
 
 export function AccessTab() {
   return (
