+++ web/features/deployments/detail/settings-tab/access/channels-section.tsx
@@ -10,7 +10,7 @@ import { SkeletonRectangle, SkeletonRow } from '@/app/components/base/skeleton'
 import { consoleQuery } from '@/service/client'
 import { DeploymentEmptyState, DeploymentNoticeState, DeploymentStateMessage } from '../../../components/empty-state'
 import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
-import { Section } from '../../common'
+import { Section } from '../../components/section'
 import { CopyPill, EndpointRow } from './common'
 import { accessSettingsQueryAtom } from './state'
 import { getUrlOrigin } from './url'
